refactor: moved util files
