+++ assembly/util/calendar.ts
@@ -1,0 +1,140 @@
+import { checkRange, floorDiv } from ".";
+
+// @ts-ignore
+@lazy
+const YEAR_MIN = -271821;
+
+// @ts-ignore
+@lazy
+const YEAR_MAX =  275760;
+
+// @ts-ignore
+@lazy
+let __null = false;
+
+// modified of
+// https://github.com/tc39/proposal-temporal/blob/49629f785eee61e9f6641452e01e995f846da3a1/polyfill/lib/ecmascript.mjs#L2157
+// @ts-ignore: decorator
+@inline
+export function leapYear(year: i32): bool {
+  return (year % 4 == 0 && year % 100 != 0) || year % 400 == 0;
+}
+
+// modified of
+// https://github.com/tc39/proposal-temporal/blob/49629f785eee61e9f6641452e01e995f846da3a1/polyfill/lib/ecmascript.mjs#L2188
+export function dayOfYear(year: i32, month: i32, day: i32): i32 {
+  const cumsumMonthDays = memory.data<u16>([
+    0,
+    31, // Jan
+    31 + 28, // Feb
+    31 + 28 + 31, // Mar
+    31 + 28 + 31 + 30, // Apr
+    31 + 28 + 31 + 30 + 31, // May
+    31 + 28 + 31 + 30 + 31 + 30, // Jun
+    31 + 28 + 31 + 30 + 31 + 30 + 31, // Jul
+    31 + 28 + 31 + 30 + 31 + 30 + 31 + 31, // Aug
+    31 + 28 + 31 + 30 + 31 + 30 + 31 + 31 + 30, // Sep
+    31 + 28 + 31 + 30 + 31 + 30 + 31 + 31 + 30 + 31, // Oct
+    31 + 28 + 31 + 30 + 31 + 30 + 31 + 31 + 30 + 31 + 30, // Nov
+  ]);
+  return (
+    day +
+    i32(load<u16>(cumsumMonthDays + ((month - 1) << 1))) +
+    i32(month >= 3 && leapYear(year))
+  );
+}
+
+// modified of
+// https://github.com/tc39/proposal-temporal/blob/49629f785eee61e9f6641452e01e995f846da3a1/polyfill/lib/ecmascript.mjs#L2164
+export function daysInMonth(year: i32, month: i32): i32 {
+  return month == 2
+    ? 28 + i32(leapYear(year))
+    : 30 + ((month + i32(month >= 8)) & 1);
+}
+
+// @ts-ignore: decorator
+@inline
+export function daysInYear(year: i32): i32 {
+  return 365 + i32(leapYear(year));
+}
+
+// Original: Disparate variation
+// Modified: TomohikoSakamoto algorithm from https://en.wikipedia.org/wiki/Determination_of_the_day_of_the_week
+// https://github.com/tc39/proposal-temporal/blob/49629f785eee61e9f6641452e01e995f846da3a1/polyfill/lib/ecmascript.mjs#L2171
+// returns day of week in range [1,7], where 7 = Sunday
+export function dayOfWeek(year: i32, month: i32, day: i32): i32 {
+  const tab = memory.data<u8>([0, 3, 2, 5, 0, 3, 5, 1, 4, 6, 2, 4]);
+
+  year -= i32(month < 3);
+  year += year / 4 - year / 100 + year / 400;
+  month = <i32>load<u8>(tab + month - 1);
+  const w = (year + month + day) % 7;
+  // Use ISO 8601 which has [1, 7] range to represent Monday-Sunday
+  return w + (w <= 0 ? 7 : 0);
+}
+
+// https://github.com/tc39/proposal-temporal/blob/49629f785eee61e9f6641452e01e995f846da3a1/polyfill/lib/ecmascript.mjs#L2667
+export function checkDateTimeRange(
+  year: i32,
+  month: i32,
+  day: i32,
+  hour: i32 = 0,
+  minute: i32 = 0,
+  second: i32 = 0,
+  millisecond: i32 = 0,
+  microsecond: i32 = 0,
+  nanosecond: i32 = 0
+): bool {
+  if (!checkRange(year, YEAR_MIN, YEAR_MAX)) {
+    return false;
+  }
+  // reject any DateTime 24 hours or more outside the Instant range
+  if ((year == YEAR_MIN && (epochFromParts(
+    year, month, day + 1, hour, minute, second,
+    millisecond, microsecond, nanosecond - 1
+  ), __null))) return false;
+
+  if ((year == YEAR_MAX && (epochFromParts(
+    year, month, day - 1, hour, minute, second,
+    millisecond, microsecond, nanosecond + 1
+  ), __null))) return false;
+
+  return true;
+}
+
+export function epochFromParts(
+  year: i32,
+  month: i32,
+  day: i32,
+  hour: i32,
+  minute: i32,
+  second: i32,
+  millisecond: i32,
+  microsecond: i32,
+  nanosecond: i32
+): i64 {
+  const millis = Date.UTC(year, month - 1, day, hour, minute, second, millisecond);
+  return millis * 1_000_000 + microsecond * 1_000 + nanosecond;
+}
+
+
+// https://github.com/tc39/proposal-temporal/blob/49629f785eee61e9f6641452e01e995f846da3a1/polyfill/lib/ecmascript.mjs#L2135
+export function weekOfYear(year: i32, month: i32, day: i32): i32 {
+  let doy = dayOfYear(year, month, day);
+  let dow = dayOfWeek(year, month, day) || 7;
+  let doj = dayOfWeek(year, 1, 1);
+
+  const week = floorDiv(doy - dow + 10, 7);
+
+  if (week < 1) {
+    return doj === 5 || (doj === 6 && leapYear(year - 1)) ? 53 : 52;
+  }
+
+  if (week === 53) {
+    if (daysInYear(year) - doy < 4 - dow) {
+      return 1;
+    }
+  }
+
+  return week;
+}
